Extract classification into own module
