+++ FossilNsmModule/ShapeCompletion.py
@@ -4,9 +4,13 @@ import ctk
 import sys
 import slicer
 import subprocess
-import json
-import glob
 from slicer.ScriptedLoadableModule import *
+
+MODULE_DIR = os.path.dirname(__file__)
+if MODULE_DIR not in sys.path:
+    sys.path.append(MODULE_DIR)
+
+from FossilNsmCommon import FossilNsmCommonWidget, FossilNsmLogic
 
 
 class ShapeCompletion(ScriptedLoadableModule):
@@ -18,67 +22,16 @@ class ShapeCompletion(ScriptedLoadableModule):
         self.parent.helpText = "A shape completion module."
 
 
-class ShapeCompletionWidget(ScriptedLoadableModuleWidget):
+class ShapeCompletionWidget(FossilNsmCommonWidget, ScriptedLoadableModuleWidget):
 
     def setup(self):
         super().setup()
 
-        # Initialize states
-        self.modelRootPath = None
-        self.inputFilePath = None
-        self.configFilePath = None
-        self.modelFilePath = None
-        self.latentCodesFilePath = None
-        self.outputFolderPath = None
-        self.referenceMeshDirectory = None
-        self.classificationMatches = []
-        self._classificationNodes = []
+        self.initializeFossilNsmState()
 
-        ShapeCompletionLogic.installDependenciesIfNeeded()
+        FossilNsmLogic.installDependenciesIfNeeded()
 
-        # Form Layout
-        inputCollapsible = ctk.ctkCollapsibleButton()
-        inputCollapsible.text = "Inputs"
-        self.layout.addWidget(inputCollapsible)
-        inputLayout = qt.QFormLayout(inputCollapsible)
-
-        # Model Root (ex: run_v44)
-        self.modelRootButton = qt.QPushButton("Select Model Root (run_vXX)...")
-        self.modelRootLabel = qt.QLabel("No model selected")
-        inputLayout.addRow("Model Root:", self.modelRootButton)
-        inputLayout.addRow("", self.modelRootLabel)
-        self.modelRootButton.connect("clicked(bool)", self.onSelectModelRoot)
-
-        # Validation output
-        self.modelChecklist = qt.QPlainTextEdit()
-        self.modelChecklist.setReadOnly(True)
-        self.modelChecklist.setFixedHeight(110)
-        inputLayout.addRow("Model Validation:", self.modelChecklist)
-
-        # Derived paths (READ ONLY)
-        self.configFileLabel = qt.QLabel("Not set")
-        self.configFileLabel.setWordWrap(True)
-        inputLayout.addRow("Config File:", self.configFileLabel)
-
-        self.modelFileLabel = qt.QLabel("Not set")
-        self.modelFileLabel.setWordWrap(True)
-        inputLayout.addRow("Model File:", self.modelFileLabel)
-
-        self.latentCodesFileLabel = qt.QLabel("Not set")
-        self.latentCodesFileLabel.setWordWrap(True)
-        inputLayout.addRow("Latent Codes File:", self.latentCodesFileLabel)
-
-        self.outputFolderLabel = qt.QLabel("Not set")
-        self.outputFolderLabel.setWordWrap(True)
-        inputLayout.addRow("Output Folder:", self.outputFolderLabel)
-
-        # Input Mesh File 
-        self.inputFileButton = qt.QPushButton("Select Input Mesh...")
-        self.inputFileButton.connect("clicked(bool)", self.onSelectInputFile)
-        self.inputFileLabel = qt.QLabel("No file selected")
-        self.inputFileLabel.setWordWrap(True)
-        inputLayout.addRow("Input Mesh:", self.inputFileButton)
-        inputLayout.addRow("", self.inputFileLabel)
+        self.addFossilNsmInputSection(self.layout)
 
         # Optimization Settings Collapsible Layout
         optimCollapsible = ctk.ctkCollapsibleButton()
@@ -179,47 +132,6 @@ class ShapeCompletionWidget(ScriptedLoadableModuleWidget):
         self.statusLog.setFixedHeight(120)
         self.layout.addWidget(self.statusLog)
 
-        # Classification is kept separate from completion. An intact mesh can be
-        # classified directly, while a completed mesh can be selected afterwards.
-        classificationCollapsible = ctk.ctkCollapsibleButton()
-        classificationCollapsible.text = "Classification: Top-5 closest meshes"
-        self.layout.addWidget(classificationCollapsible)
-        classificationLayout = qt.QFormLayout(classificationCollapsible)
-
-        self.referenceMeshesButton = qt.QPushButton("Select Reference Mesh Library...")
-        self.referenceMeshesButton.connect("clicked(bool)", self.onSelectReferenceMeshDirectory)
-        self.referenceMeshesLabel = qt.QLabel("Optional: needed to visualize returned meshes")
-        self.referenceMeshesLabel.setWordWrap(True)
-        classificationLayout.addRow("Reference Mesh Library:", self.referenceMeshesButton)
-        classificationLayout.addRow("", self.referenceMeshesLabel)
-
-        self.classificationIterationsInput = qt.QLineEdit("1000")
-        classificationLayout.addRow("Latent Optimization Iterations:", self.classificationIterationsInput)
-        self.classifyButton = qt.QPushButton("Classify Input Mesh")
-        self.classifyButton.setEnabled(False)
-        self.classifyButton.connect("clicked(bool)", self.onClassifyInputMesh)
-        classificationLayout.addRow(self.classifyButton)
-
-        self.classificationTable = qt.QTableWidget(0, 4)
-        self.classificationTable.setHorizontalHeaderLabels(["Rank", "Reference mesh", "Cosine distance", "Available"])
-        self.classificationTable.setSelectionBehavior(qt.QAbstractItemView.SelectRows)
-        self.classificationTable.setSelectionMode(qt.QAbstractItemView.SingleSelection)
-        self.classificationTable.setEditTriggers(qt.QAbstractItemView.NoEditTriggers)
-        self.classificationTable.horizontalHeader().setStretchLastSection(True)
-        classificationLayout.addRow("Top-5 matches:", self.classificationTable)
-        self.showSelectedMatchButton = qt.QPushButton("Show Selected Match")
-        self.showSelectedMatchButton.setEnabled(False)
-        self.showSelectedMatchButton.connect("clicked(bool)", self.onShowSelectedMatch)
-        self.showAllMatchesButton = qt.QPushButton("Show All Available Matches")
-        self.showAllMatchesButton.setEnabled(False)
-        self.showAllMatchesButton.connect("clicked(bool)", self.onShowAllMatches)
-        matchButtons = qt.QWidget()
-        matchButtonsLayout = qt.QHBoxLayout(matchButtons)
-        matchButtonsLayout.setContentsMargins(0, 0, 0, 0)
-        matchButtonsLayout.addWidget(self.showSelectedMatchButton)
-        matchButtonsLayout.addWidget(self.showAllMatchesButton)
-        classificationLayout.addRow(matchButtons)
-
         # Similarity Metrics Form Layout
         distanceCollapsible = ctk.ctkCollapsibleButton()
         distanceCollapsible.text = "Similarity Metrics"
@@ -270,16 +182,6 @@ class ShapeCompletionWidget(ScriptedLoadableModuleWidget):
 
         self.layout.addStretch(1)
 
-        # Update labels to reflect pre-filled defaults
-        if self.configFilePath:
-            self.configFileLabel.setText(self.configFilePath)
-        if self.modelFilePath:
-            self.modelFileLabel.setText(self.modelFilePath)
-        if self.latentCodesFilePath:
-            self.latentCodesFileLabel.setText(self.latentCodesFilePath)
-        if self.outputFolderPath:
-            self.outputFolderLabel.setText(self.outputFolderPath)
-
         self.updateRunButton()
 
     # Toggle Uncertainty Inputs
@@ -295,255 +197,8 @@ class ShapeCompletionWidget(ScriptedLoadableModuleWidget):
         is_mc = self.propagationModeCombobox.currentText == "montecarlo"
         self.nSamplesInput.setEnabled(enabled and is_mc)
 
-    # File/Folder Selection
-    def onSelectConfigFile(self):
-        path = qt.QFileDialog.getOpenFileName(
-            None, "Select Config File", "", "Config Files (*.json)"
-        )
-        if path:
-            self.configFilePath = path
-            self.configFileLabel.setText(path)
-            self.updateRunButton()
-
-    # Auto populate filepaths from model root and validate
-    def validateModelRoot(self, rootDir):
-        checks = []
-
-        def check(label, path, isDir=False):
-            ok = os.path.isdir(path) if isDir else os.path.isfile(path)
-            checks.append((label, ok, path))
-            return ok
-
-        check("model_params_config.json", os.path.join(rootDir, "model_params_config.json"))
-        check("model/ folder", os.path.join(rootDir, "model"), isDir=True)
-        check("latent_codes/ folder", os.path.join(rootDir, "latent_codes"), isDir=True)
-        # Optional output folder (not required)
-        check("shape_completion/ (optional)", os.path.join(rootDir, "shape_completion"), isDir=True)
-        return checks
-
-    # Print validated filepath checklist in UI
-    def updateChecklistUI(self, rootDir):
-        checks = self.validateModelRoot(rootDir)
-        lines = []
-        allRequiredOk = True
-
-        for label, ok, path in checks:
-            icon = "✔" if ok else "✖"
-            lines.append(f"{icon} {label}")
-            if not ok and "optional" not in label:
-                allRequiredOk = False
-
-        self.modelChecklist.setPlainText("\n".join(lines))
-        return allRequiredOk
-
-    # From specified model root, resolve other default paths
-    def resolveModelRoot(self, rootDir):
-        config = os.path.join(rootDir, "model_params_config.json")
-        modelDir = os.path.join(rootDir, "model")
-        latentDir = os.path.join(rootDir, "latent_codes")
-        outputDir = os.path.join(rootDir, "shape_completion")
-
-        # Validate required structure
-        missing = []
-        if not os.path.isfile(config):
-            missing.append("model_params_config.json")
-        if not os.path.isdir(modelDir):
-            missing.append("model/")
-        if not os.path.isdir(latentDir):
-            missing.append("latent_codes/")
-
-        if missing:
-            raise ValueError(f"Invalid model package. Missing: {missing}")
-
-        # Try to auto-detect files inside folders
-        modelFiles = sorted([f for f in os.listdir(modelDir) if f.endswith(".pth")])
-        latentFiles = sorted([f for f in os.listdir(latentDir) if f.endswith(".pth")])
-
-        if not modelFiles:
-            raise ValueError("No .pth file found in model/")
-        if not latentFiles:
-            raise ValueError("No .pth file found in latent_codes/")
-
-        modelPath = os.path.join(modelDir, modelFiles[-1])
-        latentPath = os.path.join(latentDir, latentFiles[-1])
-
-        return config, modelPath, latentPath, outputDir
-
-    # Select model root dir
-    def onSelectModelRoot(self):
-        path = qt.QFileDialog.getExistingDirectory(None, "Select Model Root Folder")
-        if not path:
-            return
-
-        # Reset paths to avoid stale paths causing problems
-        self.configFilePath = None
-        self.modelFilePath = None
-        self.latentCodesFilePath = None
-        self.outputFolderPath = None
-
-        self.modelRootPath = path
-        self.modelRootLabel.setText(path)
-
-        ok = self.updateChecklistUI(path)
-        if not ok:
-            self.onLogMessage("Model root is incomplete. Fix missing files before running.")
-            self.runButton.setEnabled(False)
-            return
-
-        config, model, latents, output = self.resolveModelRoot(path)
-
-        self.configFilePath = config
-        self.modelFilePath = model
-        self.latentCodesFilePath = latents
-        self.outputFolderPath = output
-
-        self.configFileLabel.setText(config)
-        self.modelFileLabel.setText(model)
-        self.latentCodesFileLabel.setText(latents)
-        self.outputFolderLabel.setText(output)
-
-        self.updateRunButton()
-
-    # Input mesh selector
-    def onSelectInputFile(self):
-        path = qt.QFileDialog.getOpenFileName(
-            None, "Select Input Mesh", "", "Mesh Files (*.vtk *.vtp *.stl *.obj *.ply)"
-        )
-        if not path:
-            return
-        self.inputFilePath = path
-        self.inputFileLabel.setText(path)
-        self.updateRunButton()
-
-    def onSelectReferenceMeshDirectory(self):
-        path = qt.QFileDialog.getExistingDirectory(None, "Select Reference Mesh Library")
-        if not path:
-            return
-        self.referenceMeshDirectory = path
-        self.referenceMeshesLabel.setText(path)
-        self._updateClassificationTable()
-
     def updateRunButton(self):
-        ready = bool(self.modelRootPath and self.inputFilePath)
-        self.runButton.setEnabled(ready)
-        self.classifyButton.setEnabled(ready)
-
-    def onClassifyInputMesh(self):
-        try:
-            iterations = int(self.classificationIterationsInput.text)
-            if iterations < 1:
-                raise ValueError
-        except ValueError:
-            slicer.util.errorDisplay("Latent optimization iterations must be a positive integer.")
-            return
-
-        resultDirectory = os.path.join(self.outputFolderPath, "classification")
-        os.makedirs(resultDirectory, exist_ok=True)
-        inputBase = os.path.splitext(os.path.basename(self.inputFilePath))[0]
-        self._classificationResultPath = os.path.join(resultDirectory, inputBase + "_top5.json")
-        logPath = os.path.join(resultDirectory, inputBase + "_classification.log")
-        workerScript = os.path.join(os.path.dirname(os.path.dirname(__file__)), "classification_slicer.py")
-        self._classificationLog = open(logPath, "w")
-        self._classificationLogReadPos = 0
-        self._classificationLogPath = logPath
-        self.classifyButton.setEnabled(False)
-        self.onLogMessage("Starting nearest-mesh classification (latent optimization may take several minutes)...")
-        self._classificationProcess = subprocess.Popen([
-            sys.executable, workerScript, "--config", self.configFilePath,
-            "--model", self.modelFilePath, "--latent_codes", self.latentCodesFilePath,
-            "--input_mesh", self.inputFilePath, "--output_dir", resultDirectory,
-            "--result", self._classificationResultPath, "--iterations", str(iterations),
-        ], stdout=self._classificationLog, stderr=subprocess.STDOUT)
-        self._classificationTimer = qt.QTimer()
-        self._classificationTimer.setInterval(500)
-        self._classificationTimer.timeout.connect(self._pollClassification)
-        self._classificationTimer.start()
-
-    def _pollClassification(self):
-        try:
-            with open(self._classificationLogPath, "r", errors="replace") as stream:
-                stream.seek(self._classificationLogReadPos)
-                text = stream.read()
-                self._classificationLogReadPos = stream.tell()
-                for line in text.splitlines():
-                    if line.strip():
-                        self.onLogMessage(line)
-        except FileNotFoundError:
-            pass
-        code = self._classificationProcess.poll()
-        if code is None:
-            return
-        self._classificationTimer.stop()
-        self._classificationLog.close()
-        self.classifyButton.setEnabled(True)
-        if code != 0 or not os.path.isfile(self._classificationResultPath):
-            self.onLogMessage("Classification failed (exit code {}).".format(code))
-            return
-        with open(self._classificationResultPath, encoding="utf-8") as stream:
-            self.classificationMatches = json.load(stream).get("matches", [])
-        self._updateClassificationTable()
-        self.onLogMessage("Classification complete. Top-5 matches saved to " + self._classificationResultPath)
-
-    def _referencePath(self, match):
-        if not self.referenceMeshDirectory:
-            return None
-        name = match.get("mesh_name", "")
-        direct = os.path.join(self.referenceMeshDirectory, name)
-        if os.path.isfile(direct):
-            return direct
-        found = glob.glob(os.path.join(self.referenceMeshDirectory, "**", name), recursive=True)
-        return found[0] if found else None
-
-    def _updateClassificationTable(self):
-        self.classificationTable.setRowCount(len(self.classificationMatches))
-        available = 0
-        for row, match in enumerate(self.classificationMatches):
-            meshPath = self._referencePath(match)
-            values = [str(match["rank"]), match["mesh_name"], "{:.6f}".format(match["cosine_distance"]),
-                      "Yes" if meshPath else "No — select matching library"]
-            for column, value in enumerate(values):
-                self.classificationTable.setItem(row, column, qt.QTableWidgetItem(value))
-            if meshPath:
-                available += 1
-        self.showSelectedMatchButton.setEnabled(available > 0)
-        self.showAllMatchesButton.setEnabled(available > 0)
-
-    def _clearClassificationNodes(self):
-        for node in self._classificationNodes:
-            if node and slicer.mrmlScene.IsNodePresent(node):
-                slicer.mrmlScene.RemoveNode(node)
-        self._classificationNodes = []
-
-    def _showMatch(self, match, offset=0.0):
-        meshPath = self._referencePath(match)
-        if not meshPath:
-            self.onLogMessage("Reference mesh is not available: " + match["mesh_name"])
-            return None
-        node = slicer.util.loadModel(meshPath)
-        node.SetName("Top {} — {} ({:.4f})".format(match["rank"], match["mesh_name"], match["cosine_distance"]))
-        node.CreateDefaultDisplayNodes()
-        node.GetDisplayNode().SetColor(0.2, 0.65, 0.9)
-        if offset:
-            transform = slicer.mrmlScene.AddNewNodeByClass("vtkMRMLLinearTransformNode", node.GetName() + " transform")
-            transform.GetTransformToParent().Translate(offset, 0, 0)
-            node.SetAndObserveTransformNodeID(transform.GetID())
-            self._classificationNodes.append(transform)
-        self._classificationNodes.append(node)
-        return node
-
-    def onShowSelectedMatch(self):
-        rows = self.classificationTable.selectionModel().selectedRows()
-        if not rows:
-            slicer.util.infoDisplay("Select a result row first.")
-            return
-        self._clearClassificationNodes()
-        self._showMatch(self.classificationMatches[rows[0].row()])
-
-    def onShowAllMatches(self):
-        self._clearClassificationNodes()
-        for index, match in enumerate(self.classificationMatches):
-            self._showMatch(match, offset=index * 50.0)
-        slicer.app.layoutManager().threeDWidget(0).threeDView().resetFocalPoint()
+        self.runButton.setEnabled(self.commonInputsReady())
 
     # Inference
     def onRunInference(self):
@@ -755,67 +410,5 @@ class ShapeCompletionWidget(ScriptedLoadableModuleWidget):
         self.statusLog.appendPlainText(str(message))
 
 
-class ShapeCompletionLogic(ScriptedLoadableModuleLogic):
-
-    @staticmethod
-    def installDependenciesIfNeeded():
-        # Conditionally use tiny3d (o3d) for Slicer
-        USE_TINY3D = True  # Set this to False in normal use (outside Slicer)
-
-        if USE_TINY3D:
-            try:
-                import tiny3d as o3d  # Import tiny3d as o3d for Slicer logic
-            except ImportError:
-                slicer.util.pip_install('tiny3d')
-                import tiny3d as o3d  # Try again after installing
-        else:
-            try:
-                import open3d as o3d  # Use open3d for the rest of the project
-            except ImportError:
-                slicer.util.pip_install('open3d')
-                import open3d as o3d  # Try again after installing
-
-        try:
-            import cv2
-        except ImportError:
-            slicer.util.pip_install('opencv-python')
-
-        try:
-            import nibabel
-        except ImportError:
-            slicer.util.pip_install('nibabel')
-
-        try:
-            import pymskt
-        except ImportError:
-            slicer.util.pip_install('mskt')
-
-        try:
-            import pyvista
-        except ImportError:
-            slicer.util.pip_install('pyvista')
-
-        try:
-            import pymeshfix
-        except ImportError:
-            slicer.util.pip_install('pymeshfix')
-
-        try:
-            import skimage
-        except ImportError:
-            slicer.util.pip_install('scikit-image')
-
-        try:
-            import sklearn
-        except ImportError:
-            slicer.util.pip_install('scikit-learn')
-
-        try:
-            import torch
-        except ImportError:
-            slicer.util.pip_install('torch')
-
-        try:
-            import vtk
-        except ImportError:
-            slicer.util.pip_install('vtk')
+class ShapeCompletionLogic(FossilNsmLogic):
+    pass
